+++ bin/check-snmp.rb
@@ -71,6 +71,13 @@ class CheckSNMP < Sensu::Plugin::Check::CLI
          short: '-t timeout (seconds)',
          default: '1'
 
+  option :debug,
+         short: '-D',
+         long: '--debug',
+         description: 'Enable debugging to assist with inspecting OID values / data.',
+         boolean: true,
+         default: false
+
   def run
     begin
       manager = SNMP::Manager.new(host: config[:host].to_s,
@@ -78,6 +85,10 @@ class CheckSNMP < Sensu::Plugin::Check::CLI
                                   version: config[:snmp_version].to_sym,
                                   timeout: config[:timeout].to_i)
       response = manager.get([config[:objectid].to_s])
+      if config[:debug]
+        puts 'DEBUG OUTPUT:'
+        response.each_varbind { |vb| puts vb.inspect }
+      end
     rescue SNMP::RequestTimeout
       unknown "#{config[:host]} not responding"
     rescue => e